Refactor: simplify package structure - Add a Protocol package for handling communication with the backend - Combine the types definition and methods for types into the same location
