+++ frontend/src/components/rubric-builder/RubricBuilder.tsx
@@ -15,17 +15,17 @@ import Dialog from '../util/Dialog.tsx';
 import CSVUpload from './CSVUpload.tsx';
 import Header from '../util/Header.tsx';
 import Footer from '../util/Footer.tsx';
-import { Rubric } from '../../models/types/rubric.ts';
-import createRubric from '../../models/Rubric.ts';
-import { RubricCriterion } from '../../models/types/rubricCriterion';
-import createRubricCriterion from '../../models/RubricCriterion.ts';
+import { Rubric } from '../../models/types/Rubric.ts';
+import createRubric from '../../models/types/Rubric.ts';
+import { RubricCriterion } from '../../models/types/RubricCriterion.ts';
+import createRubricCriterion from '../../models/types/RubricCriterion.ts';
 import { DndContext } from '@dnd-kit/core';
 import {
   SortableContext,
   verticalListSortingStrategy,
 } from '@dnd-kit/sortable';
-import createRating from '../../models/RubricRating.ts';
-import { BackendAPI } from '../../models/BackendRequests.ts';
+import createRating from '../../models/types/RubricRating.ts';
+import { BackendAPI } from '../../Protocol/BackendRequests.ts';
 import ModalChoiceDialog from '../util/ModalChoiceDialog.tsx';
 
 export default function RubricBuilder(): ReactElement {
