+++ frontend/src/models/types/RubricCriterion.ts
@@ -1,6 +1,15 @@
-import { RubricCriterion } from './types/rubricCriterion.ts';
-import { RubricRating } from './types/rubricRating.ts';
+import { RubricRating } from './RubricRating.ts';
 import { v4 as uuidv4 } from 'uuid';
+
+export interface RubricCriterion {
+  description: string;
+  longDescription: string;
+  ratings: RubricRating[];
+  id?: number; // id is only assigned when the rubric is saved to the database
+  points: number;
+  updatePoints: () => void;
+  key: string; // UUID for React
+}
 
 /**
  * Helper function to calculate a criterion's max point value on creation
@@ -11,7 +20,7 @@ export const calcMaxPoints = (ratings: RubricRating[]): number => {
   if (ratings.length > 0) {
     return ratings.reduce(
       (max, current) => (current.points > max.points ? current : max),
-      ratings[0]
+      ratings[0],
     ).points;
   } else {
     return 0;
@@ -26,7 +35,7 @@ export default function createRubricCriterion(
   longDescription: string = '',
   points: number = 0,
   ratings: RubricRating[] = [],
-  id: number | undefined = undefined
+  id: number | undefined = undefined,
 ): RubricCriterion {
   return {
     ratings,
