--- frontend/src/models/types/rubricRating.ts
@@ -1,9 +1,0 @@
-export interface RubricRating {
-  points: number;
-  description: string;
-  longDescription: string;
-  id?: number;
-  key: string; // UUID for React
-}
-
-export default RubricRating;
